pyproject.toml
[project]
name = "jota-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.2",
    "dj-database-url>=2.3.0",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "python-decouple>=3.8",
    "redis>=6.1.0",
]
